Fix VS2022 version17.5 multiple class in WinForm issue (#710)
Fix legacy dividebyzero exception in AMain.cs
