+++ Client/Utils/FileHelper.cs
@@ -1,0 +1,42 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Client.Utils
+{
+    public class Download
+    {
+        public FileInformation Info;
+        public long CurrentBytes;
+        public bool Completed;
+    }
+
+    public class FileInformation
+    {
+        public string FileName; //Relative.
+        public int Length, Compressed;
+        public DateTime Creation;
+
+        public FileInformation()
+        {
+
+        }
+        public FileInformation(BinaryReader reader)
+        {
+            FileName = reader.ReadString();
+            Length = reader.ReadInt32();
+            Compressed = reader.ReadInt32();
+
+            Creation = DateTime.FromBinary(reader.ReadInt64());
+        }
+        public void Save(BinaryWriter writer)
+        {
+            writer.Write(FileName);
+            writer.Write(Length);
+            writer.Write(Compressed);
+            writer.Write(Creation.ToBinary());
+        }
+    }
+}
